a roundabout identity function
The Cabal library already has a function
Distribution.PackageDescription.Parse.readPackageDescription which
parses an entire .cabal file, but writing it back out using
writePackageDescription loses all of the whitespace, reorders the lines,
and adds a few default ones, making the original file unrecognizable.
Since I want to modify the user's cabal file as unintrusively as
possible, I wrote a custom parser which keeps most of the cabal file as
raw strings, except for the dependencies. Now the only change when
printing out is the occasional addition of "-any" to dependencies which
didn't specify version information at all.

The custom parsing code isn't very pretty, but it does its job.